requirements.txt
PyYAML>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.3
Pillow>=10.4.0
